better HTTP exceptions
