+++ src/Docker/Http/Exception/TimeoutException.php
@@ -2,16 +2,12 @@
 
 namespace Docker\Http\Exception;
 
-use Docker\Exception as BaseException;
+use Docker\Http\Exception as HttpException;
 use Docker\Http\Request;
 
 /**
  * Docker\Http\Exception\TimeoutException
  */
-class TimeoutException extends BaseException
+class TimeoutException extends HttpException
 {
-    public static function fromRequest(Request $request)
-    {
-        return new self(sprintf('Timed out while waiting for a response to "%s"', $request->getRequestUri()));
-    }
 }
